TR/pid_tuning_graph/requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
future==0.18.3
iso8601==2.1.0
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0